UPDATING ERROR MESSAGES FOR PARSING (#5):
* WIP PARSE_MAP:

Adding directions to .h file.
Rewording to shorten functions to please norminette.
Adding files to Makefile

* WIP PARSING MAP CONTENT:

Still a WIP, function is still empty yet.

* UPDATE TEST MAP CHECK:

Simplifying display on test printf for map parsing.
Removing trailing \n and line: prefix.

This may cause buffer related issues.

* WIP MAP PARSING:

Adding first test cases checking floor and ceilling lines matching
patterns.

This is a WIP

* ADD FT_STRCMP:

Adding ft_strcmp to libft.

* ADD MAP CHECK CONTENT:

Adding function pointers to check for map content.
Smart solution without intensive use of if conditions.

Changes to be committed:
	modified:   include/cub3d.h
	modified:   srcs/main.c
	new file:   srcs/parsing/check_functions.c
	modified:   srcs/parsing/map_check_content.c

* ADD FT_SPLIT_F FUNCTION:

Adding a split function to use with a function instead of an int.
Check prototype to know how to use it.

Changes to be committed:
	modified:   Makefile
	new file:   ft_split_f.c
	modified:   include/libft.h

* FIX INCORRECT ARRAY DEFINITION:

Removing trailing \0 not needed for this case.

* UPDATE REWORDING FILENAMES:

Rewording filenames to be more explicit in parsing directory.

* FIX FT_SPLIT_F ADD IS_WHITESPACE:

Fixing function ft_split_f, using incorrect conditions for function
return.

Also adding function is whitespace to create a split function
that split on any whitespace given.

* ADD TOKENISER FOR MAP FLAG AND PATH:

This is a working solution with no delimiter for the map given
as input.
Now needs to be modified to stop when all textures are stored or when
map starts being drawn and trigger errors if the textures are not all
defined.

The path are not checked yet.

* ADD FT_FREE_DOUBLE_ARRAY:

Adding function to libft to entirely free a char**

* ADD FT_STRPTR_LEN:

Adding function that returns the length of a char **

* FIX WRONG ERROR MESSAGE:

Fixing wrong error message for incorrect map.
Needs to be changed for the other error messages.

* WIP ERROR HANDLING AND EXIT:

Working on a way to print the correct error message and exit during
parsing.

Still a few leaks.

Need to find a workaround repetitive functions for textures.

Now checking if the wall are files that can be opened and xpm.

* FIX FT_FREE_DOUBLE_ARRAY:

Fixing free double array now returning a void* to return NULL
So the array can be freed and set to NULL using the return.

* FIX LEAKS IN PARSING:

Working on a clean solution to exit if map is incorrect.

* FIX GET_NEXT_LINE:

Fix incorrect protection on get_next_line.

This was a major change to be done on current gnl.
It was easier to change the gnl to aweaver's.

This resulted in several functions added to libft:
	ft_strcat
	ft_strcpy
	ft_strjoin_free (frees s1)
	ft_strlen_int (returns an int instead of a size_t)
	ft_strlen_nl (counts up to \n or \0)

* UPDATE ERROR MESSAGE:

Error message now takes three args, so it is easier to use with
ft_putstr_fd. And shows with texture is causing the error.

* UPDATE ERR MSG DISPLAY USING DEFINES:

Change done to the way the messages are dealt with, now not needing
a \n in the call. The \n is automatically added by the function.

Also using defines instead of messy arbitrary numbers when returning

* ADD/FIX BIG UPDATE ON ERR HANDLING:

Massive update on several files.
Updating the map tester.
Updating the way the functions return to main.
Use of exit in some function for convenience.

Changes to be committed:
	modified:   Makefile
	modified:   include/cub3d.h
	modified:   srcs/main.c
	modified:   srcs/parsing/check_map_content.c
	modified:   srcs/parsing/check_map_name.c
	modified:   srcs/parsing/map_to_list.c
	renamed:    .test_maps.sh -> srcs/test/test_maps.sh
	deleted:    srcs/tests/.cub
	deleted:    srcs/tests/main_test_map.c
	new file:   srcs/utils/error_message.c
	modified:   srcs/utils/free_functions.c

* UPDATE .H FILE:

Moving ft_wrong_map to proper line in .h file.

* FIX TEST_MAP RULE:

Fixing test_map rule in Makefile to add the newly created funct err_msg

* FIX .GITIGNORE TEST RULE:

Test rule was ignoring test directory.

* FIX TEST_MAP.SH:

Adding newlines for easier read output.