ADD MAP GRID: (#10)
* UPDATE REMOVING ERR FROM STRUCT:

Removing error from struct. Was not needed and not used.

* FIX MAP_TO_LIST:

Fixing an issue if map had an incorrect name fd would never be
closed.

Now calling function to check name first then open.

* UPDATE FT_CHECK:

Updating ft_check no longer using fake return.
The function was originaly built in a way such that the return code
was supposedly used in upper scope.
In effect it never was. So this was changed for consistency.

Now using ft_wrong_map_exit to cleanly quit.

* UPDATE MAKEFILE:

This update is supposed to improve the libft dependencies.
Such as libft does relink if needed from the upper makefile.

* ADD MAP TO STR:

Adding map to str function. To create a tab of strings for the map
using coords [y][x]

* FIX CHECK_MAP_GRID FILE NAME:

Fixing file name.
Adding file to Makefile.

* FIX -- Revert "UPDATE MAKEFILE:"

This reverts commit 9aec0ee51a251cd602d1517567593e43243e38bd.

Previous fix resulted in incoherent makefile and multiple libft compile

* FIX FT_SKIP_HEADER:

Function to skip infos before map grid was not correctly built.
It would result in F and C lines returning as end of header if
colours had a 0 or a 1.

* ADD MAP GRID:

Adding map grid to data structure.

Needs to be thoroughly tested and check for wrong mallocs as well.

* FIX TEST RULE IN MAKEFILE:

Fixing rules test_map and vtest map.