ADD MOVEMENTS FORWARD/BACKWARD & STRAFING:
Adding movement based on N/S/E/W. These do not include calculations
based on rotation yet.
Though they check if the move is valid compared to the map.

ft_init_player had to be removed.
init player is now done during map check.

the player is now spawning on actual position in map.
each square of the map is 64pixels wide.
each step is 4px.

The map is not displayed yet but collision can be triggered nonetheless.