FIX MINIMAP:
Minimap now working with offsets on small and big maps.

Next step is adding a more detailed player icon on the minimap.

This is a combination of 5 commits.

WIP MINIMAP PRINT:

Adding minimap print, kinda working with centering.
Using for loops and not ready for norminette yet.

WIP MINIMAP:

Working on a minimap to scale it based on a define.
The ideal minimap would accept sizes that can be divided by 8.

Working on a solution to center the minimap on the player without
displaying out of boundaries if player is too close from a wall.

Close from a solution.
Working on function get player offset.
Stuck on tab size that cannot be dynamic

WIP BEFORE BRANCHING ON SQUARE MINIMAP:

Changing all the values to work with defines.
This step highlighted issues encountered with the shape of the map.

Therefore resulting in a side branch for a square map.

WIP MINIMAP LOOKS LIKE CANAL+ WITHOUT DECODER:

Yeah this kind of day at 42.

Trying hard, not succeeding... yet.

FIX MINIMAP:

Minimap now working with offsets on small and big maps.

Next step is adding a more detailed player icon on the minimap.