ADD SQUARE MINIMAP: (#18)
* UPDATE FORMAT MINIMAP DISPLAY:

Formating function to display minimap.

	modified:   include/cub3d.h
	modified:   srcs/image/minimap.c

* ADD NEW MAPS:

Adding new maps to test smaller maps.

* ADD BIG BORING MAP:

Adding big and boring map and modifying the makefile to test it.

* FIX FT_KEY_CHECKS:

Fixing key_check to remove previous safety that were orientation
dependant. Therefore not working with rotation feature.

Also moving the file to the most accurate directory.

* FIX PLAYER RENDITION:

Fix player rendition to stop any print from outstide of the map.

* FIX MINIMAP:

Minimap now working with offsets on small and big maps.

Next step is adding a more detailed player icon on the minimap.

This is a combination of 5 commits.

WIP MINIMAP PRINT:

Adding minimap print, kinda working with centering.
Using for loops and not ready for norminette yet.

WIP MINIMAP:

Working on a minimap to scale it based on a define.
The ideal minimap would accept sizes that can be divided by 8.

Working on a solution to center the minimap on the player without
displaying out of boundaries if player is too close from a wall.

Close from a solution.
Working on function get player offset.
Stuck on tab size that cannot be dynamic

WIP BEFORE BRANCHING ON SQUARE MINIMAP:

Changing all the values to work with defines.
This step highlighted issues encountered with the shape of the map.

Therefore resulting in a side branch for a square map.

WIP MINIMAP LOOKS LIKE CANAL+ WITHOUT DECODER:

Yeah this kind of day at 42.

Trying hard, not succeeding... yet.

FIX MINIMAP:

Minimap now working with offsets on small and big maps.

Next step is adding a more detailed player icon on the minimap.

* UPDATE REMOVING DEBUG PRINTFS AND COMMENTS:

Removing debuff prints from several files before merging.
Also removing defines no longer in use.