UPDATE MULTIPLE INPUT:
Now working solution for multiple key inputs with opposite
keys handled.

This is a combination of 2 commits.
This is the 1st commit message:

WIP MULTIPLE INPUTS:

This is a work in progress working solution bot not cleaned yet of
all the comments and trials and errors.

This will therefore be squashed later.

At the moment a few things are broken, first if no input are given during
the first loop the window closes. I think this is because of the
"previous_key" not being defined.

One of the other issues is that the rotation is broken and needs to be
sorted out.

Also there is a nasty flicker on the minimap that needs to be dealt
with.

This is the commit message #2:

FIX OPPOSITE KEY TRIGGERS:

Fixing a behaviour caused by pressing A then D then releasing A
causing a stop instead of continuation of D movement.
Same issue for W and S.

Also removing no longer needed comments.