ADD FIRST RENDITION OF WALL PROJECTION && FIX SOUTHWARD SEGFAULT:
This is a combination of 2 commits.

This is the 1st commit message:

WIP WALL PROJECTION:

First draft of the wall projection feature, which was
actually (surprisingly) succesfull.

To be fixed : -the wall are rendered but the image cannot be destroyed after
each rendition so the image is currently painted over in black
before reapplying pixel.

- there is a current issue with southward rotation which systematically
crashes.

- some irregularities appear on the walls

- fishbowl effect needs to be corrected

 This is the commit message #2:

FIX SOUTHWARD ROTATION SEGFAULT:

The Segcession was squashed thanks to General Staf's efforts