FIX TEXTURES:
 This is a combination of 5 commits.
 This is the 1st commit message:

ADD WALL_IMPACT CALCULATION:

This is a combination of 2 commits.

 This is the 1st commit message:

WIP SAVING IMPACT POINT:

The impact point on the wall will help render the correct
texture column.

 This is the commit message #2:

ADD WALL_IMPACT POINT CALCULATION:

The point of impact on the wall is stored in a variable in the
ray structure.
The function ft_cast_ray is split to please the norm.

 This is the commit message #2:

WIP APPLYING TEXTURES TO WALLS:

The initial wall_impact variable is scrapped because of its
uselessness.
A file apply_textures is created to gather all the textures
related functions.
The textures are indeed rendered, but appear to be shuffled.
However, they seem to differ depending on orientaion, yay.

 This is the commit message #3:

WIP TEXTURES TO WALL:

Unsuccessful attempt at solving the mosaic.

 This is the commit message #4:

WIP TEXTURES:

The textures are still KO, I push to save the debugs assistants
in case my mate needs them

 This is the commit message #5:

WIP WORK ON TEXTURES:

Getting closer but needs to go.