ADD WALL_IMPACT CALCULATION:
This is a combination of 2 commits.

 This is the 1st commit message:

WIP SAVING IMPACT POINT:

The impact point on the wall will help render the correct
texture column.

 This is the commit message #2:

ADD WALL_IMPACT POINT CALCULATION:

The point of impact on the wall is stored in a variable in the
ray structure.
The function ft_cast_ray is split to please the norm.